[UT] Separate OlapTableSinkTest UT
